backend/requirements.txt
Flask==2.1.2
Flask-SocketIO==5.3.6
python-socketio==5.8.0
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
python-dotenv==0.20.0
Flask-Cors==3.0.10
Werkzeug
PyJWT
bcrypt 